9284: if this account does not have its own info cluster, only deploy the LogsToEsInfo lambda if a remote endpoint is provided

+++ web-api/terraform/bin/deploy-account-specific.sh
@@ -19,6 +19,9 @@ if [[ "$ES_LOGS_INSTANCE_COUNT" -gt 0 ]]; then
   [[ -z "$ES_LOGS_INSTANCE_TYPE" ]] && echo "You must set ES_LOGS_INSTANCE_TYPE as an environment variable" && exit 1
   [[ -z "$LOG_SNAPSHOT_BUCKET_NAME" ]] && echo "You must set LOG_SNAPSHOT_BUCKET_NAME as an environment variable" && exit 1
   [[ -z "$NUM_DAYS_TO_KEEP_LOGS" ]] && echo "You must set NUM_DAYS_TO_KEEP_LOGS as an environment variable" && exit 1
+else
+  [[ -z "$ES_LOGS_EBS_VOLUME_SIZE_GB" ]] && ES_LOGS_EBS_VOLUME_SIZE_GB=0
+  [[ -z "$NUM_DAYS_TO_KEEP_LOGS" ]] && NUM_DAYS_TO_KEEP_LOGS=0
 fi
 
 ../../../../scripts/verify-terraform-version.sh
@@ -60,17 +63,17 @@ export TF_VAR_my_s3_state_key="$KEY"
 export TF_VAR_cognito_suffix="$COGNITO_SUFFIX"
 export TF_VAR_dawson_dev_trusted_role_arns="$DAWSON_DEV_TRUSTED_ROLE_ARNS"
 export TF_VAR_dns_domain="$EFCMS_DOMAIN"
-[[ -n "$ES_LOGS_CLUSTER_ARN" ]] && export TF_VAR_es_logs_cluster_arn="$ES_LOGS_CLUSTER_ARN"
+export TF_VAR_es_logs_cluster_arn="$ES_LOGS_CLUSTER_ARN"
 [[ -n "$ES_LOGS_CONSUMER_ACCOUNT_IDS" ]] && export TF_VAR_es_logs_consumer_account_ids="$ES_LOGS_CONSUMER_ACCOUNT_IDS"
-[[ -n "$ES_LOGS_EBS_VOLUME_SIZE_GB" ]] && export TF_VAR_es_logs_ebs_volume_size_gb="$ES_LOGS_EBS_VOLUME_SIZE_GB"
-[[ -n "$ES_LOGS_ENDPOINT" ]] && export TF_VAR_es_logs_cluster_endpoint="$ES_LOGS_ENDPOINT"
-[[ -n "$ES_LOGS_ENGINE_VERSION" ]] && export TF_VAR_es_logs_engine_version="$ES_LOGS_ENGINE_VERSION"
+export TF_VAR_es_logs_ebs_volume_size_gb="$ES_LOGS_EBS_VOLUME_SIZE_GB"
+export TF_VAR_es_logs_endpoint="$ES_LOGS_ENDPOINT"
+export TF_VAR_es_logs_engine_version="$ES_LOGS_ENGINE_VERSION"
 export TF_VAR_es_logs_instance_count="$ES_LOGS_INSTANCE_COUNT"
-[[ -n "$ES_LOGS_INSTANCE_TYPE" ]] && export TF_VAR_es_logs_instance_type="$ES_LOGS_INSTANCE_TYPE"
+export TF_VAR_es_logs_instance_type="$ES_LOGS_INSTANCE_TYPE"
 [[ -n "$LOG_GROUP_ENVIRONMENTS" ]] && export TF_VAR_log_group_environments="$LOG_GROUP_ENVIRONMENTS"
-[[ -n "$LOG_SNAPSHOT_BUCKET_NAME" ]] && export TF_VAR_log_snapshot_bucket_name="$LOG_SNAPSHOT_BUCKET_NAME"
+export TF_VAR_log_snapshot_bucket_name="$LOG_SNAPSHOT_BUCKET_NAME"
 export TF_VAR_lower_env_restore_roles="[\"arn:aws:iam::${LOWER_ENV_ACCOUNT_IDS//,/:role/restore_role_*\",\"arn:aws:iam::}:role/restore_role_*\"]"
-[[ -n "$NUM_DAYS_TO_KEEP_LOGS" ]] && export TF_VAR_number_of_days_to_keep_info_logs="$NUM_DAYS_TO_KEEP_LOGS"
+export TF_VAR_number_of_days_to_keep_info_logs="$NUM_DAYS_TO_KEEP_LOGS"
 export TF_VAR_zone_name="$DNS_DOMAIN"
 
 npm run build:assets
@@ -85,7 +88,7 @@ if [[ "$ES_LOGS_INSTANCE_COUNT" -eq 0 ]]; then
   terraform state rm 'module.kibana.opensearch_snapshot_repository.archived-logs' 2>/dev/null && echo "Removed OpenSearch snapshot repository from state" || echo "OpenSearch snapshot repository not in state"
   terraform state rm 'module.kibana.aws_cloudwatch_event_target.rotate_info_indices_daily' 2>/dev/null && echo "Removed EventBridge target from state" || echo "EventBridge target not in state"
   terraform state rm 'module.kibana.aws_cloudwatch_event_rule.every_day' 2>/dev/null && echo "Removed EventBridge rule from state" || echo "EventBridge rule not in state"
-  aws s3 rm "s3://${LOG_SNAPSHOT_BUCKET_NAME}" --recursive
+  [[ -n "$LOG_SNAPSHOT_BUCKET_NAME" ]] && aws s3 rm "s3://${LOG_SNAPSHOT_BUCKET_NAME}" --recursive
 fi
 
 terraform apply
